+++ pages/docs/install.tsx
@@ -3,7 +3,7 @@ import { Container, Grid, Typography } from "@mui/material";
 
 import Layout from "components/Layout";
 import Markdown from "components/Markdown";
-import mdContent from "assets/data/install/content.md";
+import mdContent from "assets/data/install.md";
 import useStyles from "assets/jss/components/layout";
 
 function Install(): ReactElement {
